reorgainize context; break up root and app
